removed torch save and enabled 16 bit training

+++ train_pl.py
@@ -82,11 +82,12 @@ def main():
                          callbacks=[checkpoint_callback, 
                                     EarlyStopping(monitor="validation loss", mode="min", patience=3),
                                     LearningRateMonitor(logging_interval='epoch')],
-                         log_every_n_steps=10)
+                         log_every_n_steps=10,
+                         precision=16)
     if args.checkpoint_path:
-        trainer.fit(model=model, datamodule=data_module, ckpt_path=args.checkpoint_path, precision=16)
+        trainer.fit(model=model, datamodule=data_module, ckpt_path=args.checkpoint_path)
     else:
-        trainer.fit(model=model, datamodule=data_module, precision=16)
+        trainer.fit(model=model, datamodule=data_module)
     logger.experiment.log_artifact(f"checkpoints/{RUN_ID}/", name=f'{EXPERIMENT_NAME}_model', type='model')
 
     logger.experiment.log_artifact(f"models/{RUN_ID}/", name=f'{EXPERIMENT_NAME}_model', type='model')
